Small updates to Values page (#624)
A few tweaks to improve the wording of the Values page. Edits were for clarity, consistency of tone (both increasing and reducing some directness), and also ensuring that the page better reflects the reality of working at PostHog. 

The 'Tread on toes' section still doesn't quite work for me - the title and content don't quite match, so either the value is not right or the text needs to be updated to better explain the value.