Updated performance review process (#1278)
We have adjusted the performance review process to be more 1-1 focused with your manager on future growth plans, as it felt like there was a lot of overlap with the 360 degree feedback session. This will also streamline the process a bit and make it a more efficient use of time.