Update engineering-hiring.md
Various tweaks, again left in the 'links' but not sure if they should be there. I changed the language around 'building something from scratch' as it came across as excluding those who haven't founded a company or built a side project, who are more likely to be from under-represented groups.