Update hiring process (#1394)
* Update hiring process

* Create Engineering-hiring.md

* Create design-hiring.md

* Create operations-hiring.md

* Rename Engineering-hiring.md to engineering-hiring.md

* Create exec-hiring.md

* Update sidebar

Not sure I was able to get the logic right

* Update sidebars.json

* Update design-hiring.md

Made various tweaks - should there be the 'link' text left in the SuperDay process? I left it in for now, but not sure what the purpose is.

* Update engineering-hiring.md

Various tweaks, again left in the 'links' but not sure if they should be there. I changed the language around 'building something from scratch' as it came across as excluding those who haven't founded a company or built a side project, who are more likely to be from under-represented groups.

* Update exec-hiring.md

I made some changes, but one thing worth noting on this (and other pages) is sometimes the language flips between candidate-facing and internally-facing. We should be consistent.

* Update operations-hiring.md

* Update hiring-process.md

* Update sidebars.json

* "Hiring process" → "Hiring Process"

Co-authored-by: Charles Cook <70321811+piemets@users.noreply.github.com>
Co-authored-by: Neil Kakkar <neilkakkar@gmail.com>
Co-authored-by: Michael Matloka <dev@twixes.com>