Increase coworking budget (#1402)
* Increase coworking budget

We currently offer $200 a month for coworking spaces or café working. With life slowly returning to a pre-covid state, a few people flagged that most coworking spaces are above the $200 mark. 

The 2021 Desk price index shows the average cost per hotdesk per month. For cities/countries where we have at least 2 people:
- London: $340 (WeWork all Access is £279+VAT = £324)
- NYC: $387
- SF: $387
- Seattle: $313
- US-wide: $277
- UK-wide: £266

Based on that, I’d suggest to increase the budget to $400 <> £325 <> €300

* clarified policy

* Add safe redirects

Co-authored-by: James Hawkins <47497682+jamesefhawkins@users.noreply.github.com>
Co-authored-by: PostHog <hey@posthog.com>