Copyedit STYLEGUIDE.md
This is a WIP of suggestions for the style guide.

I kept the same overall structure, but changed some of the headings and edited much of the body copy for clarity and brevity.

I took out a few things like the discussion of _why_ we spell "third-party cookies" the way we do. It's simpler to just give the example.

I removed one heading at the end: "Avoid repetition of terms". I can think of dozens of recommendations like this. Instead of throwing them all into a style guide, I suggest addressing them though the copyediting process, and then adding recommendations for whichever ones come up frequently.

And I added one new heading at the end: "Do not fear the copyeditor". My preference is to welcome all documentation contributions, perfect or imperfect, and we can serve as copyeditors to help make those contributions most valuable.