Wrote an opinionated marketing manifesto
We need to get more opinionated about how we run marketing at PostHog, as we have so far focused purely on tactics and shipping fast (which is ok!).  

This is an attempt to write a marketing manifesto that outlines what we stand for as a company, and how we do and don't do business. I think it's important that we are just as opinionated here as we are about product, engineering, design and people - marketing shouldn't be some commercial afterthought we stick on at the end.  

Thoughts/opinions welcome!

(Thank you to @mariusandra @kpthatsme @marcushyett-ph for feedback on the initial rough draft of this)