Minor product page tweaks
Two minor changes to the copy on the new /product page.

Firstly wanted to give a slightly clearer answer to the 'What is PostHog?' question, as we currently don't use the word 'analytics' until right at the end and it reads a little vague.

Then simplified the heatmap text to make it more direct and shorter.