Create Sessions page in the Docs
Since we removed Sessions from the app, we have linked 'Sessions' in the Docs sidebar to a blog post explaining why we did this. 

I think a better solution here is to have a Docs page that says Sessions has been removed and then link to the blog post from that, so that the experience is consistent for someone browsing the Docs, as it's jarring to be taken away to a completely different section of the site.