docs(aws): add details for ALB configuration for geoip to work
There are details for L4 load balancers to ensure that we have source
information. However, when using a L7 load balancer, we do not need to
specify to use [proxy
protocol](https://www.haproxy.org/download/1.8/doc/proxy-protocol.txt)
as we'll already have X-Forwarded-* headers, which includes
X-Forwarded-For, a header django will use for determining source ip.

Will be relevant when
https://github.com/PostHog/charts-clickhouse/pull/349 is merged.