Add advice on time spent training others
One of the outcomes of the session on diversity at the Iceland offsite was to clarify that time spent coaching at organisations that aim to increase diversity in tech is (within reason) part of your work commitments.

It's a common failure mode to have employees from a "diverse background" spend time working to increase diversity in tech without giving them paid time to do that. As a result, they're being held to the same standard as others without the same amount of time to unwind as their peers.