OS clarifications for Developing Locally Guide
Instructions specific to Debian-based Linux distributions are labelled using the term Linux only. One could argue against that term usage but at least it should clarify that this is not for all distributions and the appropriate commands for other distributions are simply an exercise for the reader.
There was also inconsistent usage of apt versus apt-get, the former appears to be preferred for end user usage now as per https://wiki.debian.org/DebianPackageManagement#Package_Management_with_apt-get_and_dpkg.
Finally, a typo fix and a clarification as to one command that is intended for macOS only.