Improve homepage hero sub-header
Inspired by this thread - I think our sub-header could do a lot more work for us. I've suggested an alternative which tells people more about what PostHog does. If we are trying to define Product OS as a new term, we need something super clear to back it up. 

It should be really obvious to a reader what PostHog is if all they read is our title and subtitle. I know we then have a carousel below, but a) the first text people read is the most important, and b) this matters for mobile.