docs: Add podman and log mentions in developing-locally (#4965)
* docs: Add podman and log mentions in developing-locally

Podman and podman-compose are valid alternatives to Docker and docker
compose respectively. This commit adds mentions to Podman where
relevant informing developers they may choose it as an alternative that
doesn't require root privileges.

Moreover, I added a small mention about ClickHouse logs as the
container does not log to stdin, but to a local file. Knowing where
these log files are can be useful when debugging setup issues.

* docs: Tune references to Podman in develop-locally.md

1. More specifically defined the Podman version that can use a host
machine's /etc/hosts.

2. Removed a paragraph mentioning Podman as an alternative to Docker,
as users of Podman should already be aware of this, and it keeps the
docs cleaner.