Add SOC 2 link to the Docs sidebar (#4935)
Imperfect solution as this links to the Handbook, but for now this at least surfaces some SOC 2 info in the Docs as well (until we get certified at least). It may be that SOC 2 should only live in the Handbook anyway, as it covers internal processes rather than specific guidance about how to use PostHog. 'Using PostHog in a SOC 2 compliant way' doesn't really make sense as a concept anyway.