Update Self Host page in the Docs (#5811)
Changed intro to put hobbyists first as that is the target user for the open source product (not people doing proof of concept). 

Removed the reference to PostHog Cloud as it felt patronizing - I would be annoyed if I came looking for this info and was getting aggressively pushed to the Cloud product.