Separate ops and people in Handbook sidebar
We want to add more ops content to the handbook, but this will make the ops & people section really long. I've split this into two sections - one for people stuff, which is mostly for the wider team, and one for ops/finance which is more runbooks for the ops team only.