Update Marketing Q3 objective and goals
Updated based on a conversation with Lior just now. Our current objective is poorly written and it's hard to tell whether or not our work is definitively contributing to that objective or not. With everything we work on, we should be able to easily say 'yes this action contributes to that objective' or 'no it doesn't, work on something else.' 

I've picked the objective that Lior and I felt most closely matches what it is we're actually trying to achieve this quarter. 

I also updated ownership of the goals and tweaked a bit based on the past couple of sprints to reflect what people are actually working on.