Add section on hierarchy/'approvals' to the handbook
Thought it would be useful to articulate how we actually approach org structure at PostHog - ie minimal hierarchy, default to _not_ going through the exec team to get stuff done. 

I think as we grow, it'll be important to not just default to traditional 'running things up and down the org to get stuff done'.