docs: Align example JSON of identify with caller
The first param to identify is to supply an id a user, in this case `distinct_id` is used as an example of a random string. But the example to the right shows `123` as the corresponding ID. I'd thought that the example would be better showcased by actually displaying the concrete id as shown in the example.

Secondly, the `createdAt` property seems to be pointing to two different years? The caller on the left is supplying 2023 as `createdAt` while the example JSON is showing `2019-01-01`.