docs: is:inline in Astro (#8273)
The user must include the `is:inline` directive or they will be confronted errors such as

`Property 'posthog' does not exist on type 'Window & typeof globalThis'.`

and

`An expression of type 'void' cannot be tested for truthiness.`

which will break builds. This includes the `is:inline` directive to fix this problem.