Extreme densification of content
The idea is that instead of writing an insanely long post about all the
different features available, it's a short bullet list that gives a
clear, concise, dense example of what these are for.

Each bullet point can have a show, don't tell approach with a togglable
visual and a link to find out more.

If this ends up being the right approach, I can put the work into
generating the screenshots and working on the styles, but for now I'm
trying to make sure we get the content right first.