Improved expandable details approach
This feels like a pretty big improvement to the previous style overrides.
I had attempted the previous version as a means of using a bullet list and
then having inline clickable emojis as the toggle. But it was pretty
unintuitive (we had to have copy that said "Click the 👀 emoji").

The other problem with that approach was the mix of HTML and Markdown meant
having to put it all on one line, and would've broken our screenshots that
can have light mode and dark mode.

This new approach cleans things up quite a bit, and it also gives us the
ability to add some further content in the details sections to add optional
explanations for what we want to show in the visuals.

Having said all that, the old version just looked like shit and the new
version looks way better.