Minor changes to JS install snippet copy
This change was based on a conversation I had with Andy about the Install
PostHog page. There are instructions for installing the JS snippet that are
critically important, but the ES5 section is likely unnecessary for most users.
So the idea was to move it into a `details` element that doesn't need to be
expanded but conceals the extra cruft for most people.

This commit includes:

1. Moving the ES5 support into a `details` element.
2. Add copy to the reverse proxy header to mention that it's optional, but recommended.
3. Adjust the CSS styles for the little SVG chevron icon so that it supports changing colors for dark mode.