fix: Avoid overlap between navbar and directional button
On some screen sizes once we have an overflow in the amount of items we display in our navbar, we display a button to allow us to scroll between the items. It just so happens, however, that on some screen sizes there's a terrible overlap between the text and the button. We're now giving much more space to the button and removing that overlap.