Deleted speculative application reference and link within "Open Roles" section
A friend found that this link redirected to the /careers/ page. 

I couldn't find a working link to the /careers/speculative-application page via site search, the Handbook or Google. The only reference in the handbook is here: https://posthog.com/handbook/people/hiring-process#application

I also lack Ashby permissions to see if there is an open role for speculative applications. 

So, I deduced that we may no longer be accepting them. Hence the deletion.

This is baby's first PostHog PR, though, so if the link needs to be fixed, feel free to reject or change. Cheers.