fix: Fix link on icon to the side of feature flags support
Our SDKs have a small info icon to the right of each of the features linking to the docs. The feature flags docs was a dead link so I've fixed it. I've also modernized some other links to stop relying on `vercel.json` to get to the right place - faster navigation.