Add support for resizing AppWindows from the left (#12814)
* Add support for resizing AppWindows from the left

Fairly straightforward. The only tricky bit is a "left resize" is actually a "resize and move" operation to keep the right edge correctly aligned. The motion animation being applied to size/position caused the window to jump around during a resize drag from the left, which has to be disabled.

Core changes:
- Add a piece of state to track if the drag operation started on the left side of the window.
- Add a helper method `handleResize` to centralize the resize logic and apply a position shift as required.
- Add new drag helpers for the left edge and bottom left corner.

An argument could be made to factor out a drag handle component to avoid some of the repetition here, particularly with configuring the motion div and tailwind classes. Let me know if you'd like me to pull that out.

Resolve #12776

* Remove unnecessary flag, update names to be more explicit.
- There is already a `handleResize` declared inside a useEffect callback. This didn't break anything (the callback definition shadowed mine), but it's confusing and unnecessary. I updated mine to `handleDragResize`, which is more descriptive anyway.
- The drag handler for resizing took a flag that specified whether the drag originated from the left side. This is redundant with the leftDragResizing state, so I removed it.

---------

Co-authored-by: Ian G McDowell <ianthedev@iangmcdowell.com>