Added line 350 - debugging interview.mdx (#13189)
* Added line 350 - debugging interview.mdx

Added a note on the debugging interview (line 350) 
As we grow, we find the need to hire engineers who are comfortable working with existing codebases to be increasingly fundamental. During Superday, Product Engineering candidates will also have a 45-minute debugging session. There is nothing to prepare in advance for this; you'll work with your interviewer in a pairing session to get through a bunch of bugs! It is a demanding day of work.

* Update contents/handbook/people/hiring-process/index.mdx

---------

Co-authored-by: Raquel Smith <raquelmsmith@users.noreply.github.com>