Update equity refreshes to be annual (#13112)
* Update equity refreshes to be annual 

We are introducing equity refreshes to be annual now so that people can continually be vesting more stock and feel the growth of that on an ongoing basis. 

This will be forward facing, so it will not effect those people who have already received (or are about to receive their 4 year refresh). This is roughly anybody who started in 2023 or earlier, for those people we are working out a process for this but it will likely mean receiving your first annual refresh to coincide with your final year of vesting in your second grant (or something that creates parity with the new system). This is a relatively small group of people (<20) so we will deal with these people directly. 

If you started after 2023 but have been here for > 1 year you will likely receive an initial refresher grant that makes up for the year(s) you did not receive one, this is also a small group of people (~15) so we will deal with you directly

* Update contents/handbook/people/compensation.mdx

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>

* Update compensation.mdx

---------

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>