Update spending-money.md
The wording 'exceptionally tall' is vague and it makes sense to put an actual number to it. Typically companies adhere to 6.5 foot or 2 meters 

Why this lenght? This length used to be an airline standard for a different seat, back when emergency exits were the same price as another seat. (At least in the Netherlands)