SUPERDAY Update foundation-check.md
It seems like most of the things that this basic account checklist/foundation checklist are data that we have access to.  We should build a dashboard that shows this checklist and the data that corresponds to each checklist and leave this page as a description of what and why we're checking.