chore: Avoid running postinstall scripts (#13830)
* chore: Avoid running postinstall scripts

These are a security risk, let's not run them if not needed

This might not work because yarn 1 doesn't let us choose what dependencies are allowed to run postinstall scripts. We'll need to migrate to pnpm if we fail to build the app with these changes.

* chore: Migrate to pnpm

yarn is very hard to make it safe after today's supply chain attack, let's use pnpm - same as our main repo

* feat: Add missing dep

This worked when using yarn because a dependency installed this, but this stopped working when using pnpm because it installs files in different places

* chore: Add existing dependencies to list

* fix: Update browserslist

* fix: Fix incorrect advice

* feat: Add all missing deps to package.json

We were previously using yarn and on yarn you can require transitive dependencies because files are installed like this

```
node_modules
  | node
  | posthog-js
  | posthog-js-dependency-i-can-require
```

on pnpm, on the other hand, they're nested, which means we cannot require dependencies we didn't install - which is much saner

This meant we had to install more packages than before to get this to work

* chore: add axios

* chore: sort package.json

* chore: Update pnpm lock