feat: Implement product interest tracking for onboarding
Added functionality to track user interest in products through landing pages and documentation. This includes the creation of a new cookie to store product slugs for personalized onboarding experiences. Key changes include:

- New utility functions in `src/lib/productInterest.ts` for managing product interest cookies.
- A custom hook `useProductInterest` to facilitate tracking in React components.
- Integration of the tracking mechanism in relevant components and templates.

We haven't implemented the changes on the main app yet, this is mostly a POC to see if we can easily track interest - hence why the cookie lasts so little.